Rename Widget to fit previous one
